feat: iOS | Edit Picture with URIs (#102)
Update `OSCameraLib.xcframework` to use the latest version that includes the new `editPicture` for URIs.
Add new `editURIPicture` to `OSCamera`. This is method enables picture editing from an URI.
Create a `OSCAMREditPictureParameters` struct to receive the parameters from the JS API. This also allows the opportunity to split all Parameter structs into separate files, making it easier to manage.